Move Subclass Registry to it's own file
